pyproject.toml
[tool.poetry]
name = "orjson-benchmark"
version = "0.1.0"
description = ""
authors = ["Dennis Fuchs <aaa@aaa.aaa>"]

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
matplotlib = "^3.5.1"
tabulate = "^0.8.9"
pytest-benchmark = "^3.4.1"
pytest-random-order = "^1.0.4"
memory-profiler = "^0.60.0"
tox = "^3.24.5"
orjson = "^3.6.7"
python-rapidjson = "^1.6"
simplejson = "^3.17.6"
ujson = "^5.1.0"
isort = "^5.10.1"
autoflake = "^1.4"
black = "^22.1.0"

[tool.poetry.scripts]
graph-boxplot = "orjson_benchmark.graph:box"
graph-tabulate = "orjson_benchmark.graph:tab"
test_dataclass = "orjson_benchmark.pydataclass:test_dataclass"
test_numpy = "orjson_benchmark.pynumpy:test_numpy"
test_memory = "orjson_benchmark.pymem:test_mem"
test_nonstr = "orjson_benchmark.pynonstr:test_nonstr"
test_sort = "orjson_benchmark.pysort:test_sort"
test_indent = "orjson_benchmark.pyindent:test_indent"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
